refactory: refactory

+++ src/infra/controllers/UserController.ts
@@ -5,9 +5,12 @@ import { CreateUser } from "application/useCases/CreateUser";
 import { DeleteUser } from "application/useCases/DeleteUser";
 import { UpdateUser } from "application/useCases/UpdateUser";
 import { User } from "domain/entities/User";
+import { NotAuthorized } from "domain/errors/NotAuthorized";
 import { Server } from "infra/http/Server";
+import { Auth } from "infra/security/Auth";
 import { Authenticator } from "infra/security/Authenticator";
 import { Encrypt } from "infra/security/Encrypt";
+import { GateAdapter } from "infra/security/GateAdapter";
 import { Identifier } from "infra/security/Identifier";
 
 export class UserController {
@@ -22,6 +25,8 @@ export class UserController {
   ) {
     this.server.post("/user", async (req, res) => {
       try {
+        const userData = req.body as Omit<User.Props, "id">;
+
         const createUser = new CreateUser(
           this.identifier,
           this.encrypt,
@@ -31,9 +36,7 @@ export class UserController {
           this.categoryRepository
         );
 
-        const user = await createUser.execute(
-          req.body as Omit<User.Props, "id">
-        );
+        const user = await createUser.execute(userData);
         res.json(user).end();
       } catch (err) {
         res.status(400).json(err.message);
@@ -42,6 +45,9 @@ export class UserController {
 
     this.server.put("/user", async (req, res) => {
       try {
+        if (GateAdapter.allows("JUST_ADM")) {
+          throw new NotAuthorized();
+        }
         const updateUser = new UpdateUser(this.userRepository, this.encrypt);
         const user = await updateUser.execute(req.body as Partial<User.Props>);
         res.json(user).end();
@@ -52,8 +58,11 @@ export class UserController {
 
     this.server.delete("/user", async (req, res) => {
       try {
+        if (!GateAdapter.allows("JUST_ADM")) {
+          throw new NotAuthorized();
+        }
         const deleteUser = new DeleteUser(this.userRepository);
-        await deleteUser.execute(req.body.id);
+        await deleteUser.execute(req.body.id ? req.body.id : "");
         res.end();
       } catch (err) {
         res.status(400).json(err.message);
@@ -71,6 +80,9 @@ export class UserController {
 
     this.server.get("/all_users", async (req, res) => {
       try {
+        if (!GateAdapter.allows("JUST_ADM")) {
+          throw new NotAuthorized();
+        }
         const users = await this.userRepository.list();
         res.json(users).end();
       } catch (err) {
